% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spatial_fun.R
\name{get_spatial}
\alias{get_spatial}
\title{Easily Collection of Digital Spatial Data}
\usage{
get_spatial(
  digital.raster = NULL,
  which.raster.number = NULL,
  lat = NULL,
  lng = NULL,
  env.dataframe = NULL,
  env.id = NULL,
  name.feature = NULL,
  merge = TRUE
)
}
\arguments{
\item{digital.raster}{RasterStack or SpatRaster file}

\item{which.raster.number}{numeric, if digital.raster is RasterStack this number denotes which raster to be used}

\item{lat}{vector (numeric). Latitude values of the site/environment (e.g. -13.05) in WGS84.}

\item{lng}{vector (numeric). Longitude values site/environment (e.g. -56.05) in WGS84.}

\item{env.dataframe}{data.frame}

\item{env.id}{vector (character or level). Identifimessageion of the site/environment (e.g. Piracicaba01).}

\item{name.feature}{vector (character)}

\item{merge}{logical}
}
\description{
Extracts point feature esitmations from large digital image files
}
\details{
==================================================================================================
Title.    : Collecting point estimates from raster files
Author.   : G Costa-Neto
Created at: 2023-11-15
Updated at: 2023-11-15
Previous versions: EnvRtype::extract_GIS()
Current Version: 0.0.1

get_spatial()
==================================================================================================
}
\examples{
\dontrun{
## Elevation data for a given site
reference_table = data.frame(env = 'NM',lat = -13.05, lng =   -56.05)
get_spatial(env.dataframe = reference_table,
digital.raster = envirotypeR::SRTM_elevation,
lat = 'lat',lng = 'lng',env.id = 'env',
name.feature = 'Elevation' )
}
}
\author{
Germano Martins Costa Neto
}
