% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_synthetic_fun.R
\name{process_synthetic}
\alias{process_synthetic}
\title{Creates Orthogonal Vectors from Environmental Features.}
\usage{
process_synthetic(
  env.dataframe,
  missing.rate = 0.05,
  n.synthetic = 15,
  digits = 5,
  impute.missing = TRUE,
  sample.imp = 5,
  return_raw = FALSE
)
}
\arguments{
\item{env.dataframe}{(data.frame) with row names representing environments and column names as environmental variables. It accepts missing values.}

\item{missing.rate}{(numeric) the minimum missing rate accepted, default is 0.05.}

\item{n.synthetic}{(numeric) number of orthogonal vectors (Moran's eigenvectors) representing synthetic environmental traits to be created.}

\item{digits}{(numeric) number of digits in the output matrices.}

\item{impute.missing}{(boolean) if TRUE, an imputation approach is conducted to impute missing data.}

\item{sample.imp}{(numeric) Number of environments to be sampled and used for creating a training set for imputation when dealing with large datasets.}

\item{return_raw}{(boolean) if TRUE, returns a FactoMineR object to be used for clustering.}
}
\value{
A list of objects containing the following elements:
\itemize{
  \item{environmental.features}{ A data frame containing the original environmental variables and the synthetic environmental traits.}
  \item{synthetic.environmental.traits}{ A data frame containing the synthetic environmental traits (eigenvectors).}
  \item{variance.explained.by.eigen}{ A data frame with the eigenvalues, explained variance, and cumulative explained variance for each synthetic trait.}
  \item{variable.components}{ A data frame with the variable components (coordinates) for each synthetic trait.}
  \item{variable.contribution}{ A data frame with the variable contributions for each synthetic trait.}
  \item{variable.correlation}{ A data frame with the variable correlations for each synthetic trait.}
  \item{raw_PC_object}{(Optional) The FactoMineR PCA object, returned only if `return_raw` is TRUE.}
}
}
\description{
Feature Imputation and Computation of Environmental Eigenvectors from raw envirotyping data.
}
\examples{
\dontrun{
## using a matrix of features pre-created (W_matrix)
W_processed <- process_synthetic(env.dataframe = envirotypeR::W_matrix_G2F)
lenght(W_processed)
}


}
\author{
Germano Costa Neto
}
