% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRTM_elevation.R
\docType{data}
\name{SRTM_elevation}
\alias{SRTM_elevation}
\title{Digital Elevation - Shuttle Radar Topography Mission - SRTM (meters above sea level)}
\format{
A .tif raster file
}
\usage{
data(SRTM_elevation)
}
\description{
The Shuttle Radar Topography Mission (SRTM) was flown aboard the space shuttle Endeavour February 11-22, 2000.
The National Aeronautics and Space Administration (NASA) and the National Geospatial-Intelligence Agency (NGA) participated
in an international project to acquire radar data which were used to create the first near-global set of land elevations.
}
\examples{
\dontrun{
data(SRTM_elevation)
require(raster)
plot(SRTM_elevation )
}
}
\keyword{datasets}
