% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WC_Bioclimate_data.R
\docType{data}
\name{WC_Bioclimate}
\alias{WC_Bioclimate}
\title{Wordclim Bioclimatic Variables}
\format{
A RasterStack file
}
\usage{
data(WC_Bioclimate)
}
\description{
A collection of 19 bioclimatic variables from WorldClim. Check <https://www.worldclim.org/data/bioclim.html>
}
\examples{
\dontrun{
data(WC_Bioclimate)
require(raster)
plot(WC_Bioclimate[[1]]) # plot the fister layer
names(WC_Bioclimate)
}
}
\references{
Fick, S.E. and R.J. Hijmans. _WorldClim 2: new 1km spatial resolution climate surfaces for global land areas_. International Journal of Climatology 37 (12): 4302-4315.
}
\keyword{datasets}
