% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_frequency_fun.R
\name{fit_frequency}
\alias{fit_frequency}
\title{Estimates the potential environmental types and computs its frequency}
\usage{
fit_frequency(
  env.data,
  env.id = "env",
  var.id,
  day.id = "daysFromStart",
  time.unit = "f",
  interaction.var = NULL,
  as.frequency = TRUE,
  quality.control = TRUE,
  rm.duplicates = TRUE,
  rm.lowMFF = TRUE,
  duplicate.rate = TRUE,
  digits = 3,
  percentile = 0.05,
  print.duplicates = FALSE,
  seed = 9182
)
}
\description{
Estimates the potential environmental types and computs its frequency
}
