% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_synthetic_fun.R
\name{process_synthetic}
\alias{process_synthetic}
\title{Feature Imputation and Computation of Environmental Eigenvectors from raw envirotyping data}
\usage{
process_synthetic(
  env.dataframe,
  missing.rate = 0.05,
  n.synthetic = 15,
  digits = 5,
  impute.missing = TRUE,
  sample.imp = 5
)
}
\description{
==================================================================================================
Title.    : Feature Imputation and Computation of Environmental Eigenvectors
Author.   : G Costa-Neto
Created at: 2023-11-27
Updated at: 2023-12-05 (envirotypeR version 0.1.3)
Current Version: 0.1.1 (envirotypeR)
}
\details{
process_synthetic(), based on nasapower::get_power() from Sparks et al 20218
==================================================================================================
}
